Added dev tools and refactored code to ApiBusyIndicator component.
