+++ web/src/App.tsx
@@ -1,27 +1,10 @@
-import React from 'react';
-import { useQuery } from 'react-query'
-import Badge from 'react-bootstrap/Badge'
+import * as React from 'react';
+import { ApiStatusIndicator } from './features/apiStatusIndicator';
 import logo from './logo.svg';
 import './App.css';
 
 function App() {
-  const isHealthyQuery = useQuery('healthCheck', () =>     
-    fetch('health-check')
-      .then(response => {
-        if (response.ok) {
-          return response.json();
-        }
 
-        return false;
-      })
-      .then(json => json.isHealthy)
-      .catch(e => {
-        console.log(e);
-        return false;
-      })
-  , { refetchInterval: 500 });
-
-  const isHealthy = isHealthyQuery?.data || false;
 
   return (
     <div className="App">
@@ -38,8 +21,7 @@ function App() {
         >
           Learn React
         </a>
-
-        <Badge variant={isHealthy ? 'success' : 'danger'}>{isHealthy ? "API is healthy": "API is not healthy"}</Badge>
+        <ApiStatusIndicator/>
       </header>
     </div>
   );
