+++ web/src/features/apiStatusIndicator/index.tsx
@@ -1,0 +1,1 @@
+export * from './ApiStatusIndicator';
